Add in-browser channel wallet (Regev/plonky3) + CLI companion
Browser wallet for in-channel send/receive on the Regev channel model:
- src/wallet_core.rs: target-independent core (SPHINCS+ + Regev keys,
  genesis, build/verify in-channel transfer with the E-1 STARK, N-of-N
  co-signing, balance decryption). Verifies REAL SLH-DSA member signatures
  (crypto_sign_verify) and rebuilds E-1 statements from authenticated state.
- src/wasm_wallet.rs: wasm-bindgen session entry points (keygen, genesis
  contribution, sign/import, balance, send, cosign, finalize). Secrets stay
  in the wasm session; Production STARK level.
- src/bin/channel_member.rs: CLI companion running the other members.
- wallet.html + wallet-worker.js: standalone browser UI; multithreaded
  proving via wasm-bindgen-rayon. build-wallet-wasm.sh builds the pkg
  (cargo rustc --crate-type cdylib + wasm-bindgen + regev-parallel) so the
  crate stays rlib-only for native bins.
- wallet-e2e.js: Playwright + CLI end-to-end (browser sends 7, receives 5).
- tests/wallet_core_e2e.rs, tests/verify_wasm_proof.rs.

Supporting changes:
- Cargo.toml: bump regev_plonky3 to 3b17b8e (canonical Poseidon2 constants;
  fixes wasm<->native proof portability). getrandom 0.4 wasm_js feature for
  the rand 0.10 path on wasm. Off-by-default legacy_wasm_demo feature.
- src/lib.rs: add wallet_core/wasm_wallet; gate the stale wasm_demo.
- src/constants.rs: MAX_NUM_CHANNELS usize->u64 (wasm32 overflow).

Threat model + lessons in tasks/wallet-*.md. BREAKING: keys/proofs from the
pre-3b17b8e Regev config must be regenerated.

+++ src/constants.rs
@@ -11,7 +11,10 @@ pub const SEND_TREE_HEIGHT: usize = 32;
 // `channel_id` ALONE (key_id lives only in the channel layer). So the base channel id is 32 bits
 // and the channel tree is indexed by channel_id.
 pub const CHANNEL_TREE_HEIGHT: usize = CHANNEL_ID_BITS;
-pub const MAX_NUM_CHANNELS: usize = 1usize << CHANNEL_ID_BITS;
+// `u64`, not `usize`: `1 << 32` overflows the 32-bit `usize` on the wasm32 target (fine on 64-bit
+// native). This constant is informational (channel-id space size) and unused elsewhere, so widening
+// the type is value-preserving on native and portable to wasm.
+pub const MAX_NUM_CHANNELS: u64 = 1u64 << CHANNEL_ID_BITS;
 
 // Private State
 pub const ASSET_TREE_HEIGHT: usize = TOKEN_INDEX_BITS;
